+++ test_basic.sh
@@ -3,4 +3,7 @@
 # ensure we bail on errors
 set -e
 
+# update the PYTHONPATH
+. ./updatePyPath.sh
+
 ./docs/development/run_tests.py ./docs/examples/*.ipynb ./docs/user_guide/*.ipynb ./docs/test/*